fixed sorting algorithm
the sorting algorithm is fully functional. currently it only works in a
test case i created, but by reading in the files in the 20 files created
by the parsing method, it is expected to pose no problems. this may be
the last release of this program due to the client proposing different
requirements for calculating the noise