This is replacing the noise calculator
This is the Iperf Noise Calculator 0.0.1 alpha build. This program is
not complete. an algoritym still neeeds to be created to pass an array
to charts4j which utilizes the google api to plot our graphs. Further
changes need to be made for taking in values in the beginning of the
program. A bash script is ran on the Iperf data, then a regex needs to
be implemented into the program to pull our values out of output from
the bash script.